+++ lab_8/src/bot/handlers/fetch.py
@@ -1,62 +1,12 @@
 from aiogram import Router
 from aiogram.types import Message
 from aiogram.filters import Command
-import aiohttp
-import datetime
 import logging
+
+from src.utils.github import get_github_user, prettify_user_info
 
 
 router = Router()
-
-
-def prettify_user_info(user_info: dict) -> str:
-    # html injection ???   
-    prettified = ""
-
-    site_admin = user_info["site_admin"]
-    if site_admin:
-        prettified += "✨ADMIN\n"
-
-    prettified += f"<b>✏️Login (username):</b> <code>{user_info['login']}</code>\n"
-
-    name = user_info["name"]
-    if name is not None:
-        prettified += f"<b>🤵Name:</b> <code>{name}</code>\n"
-    
-    bio = user_info["bio"]
-    if bio is not None:
-        prettified += f"<b>📒Bio:</b>\n<blockquote>{bio}</blockquote>\n\n"
-    
-    company = user_info["company"]
-    if company is not None:
-        prettified += f"<b>💼Company:</b> <code>{company}</code>\n"
-    
-    email = user_info["email"]
-    if email is not None:
-        prettified += f"<b>📧Email:</n> {email}"
-    
-    created_date = datetime.datetime.strptime(user_info['created_at'], "%Y-%m-%dT%H:%M:%SZ")
-    created_date_string = datetime.datetime.strftime(created_date, "%Y/%m/%d")
-    prettified += f"<b>🕒Created at:</b> <code>{created_date_string}</code>\n\n"
-
-    prettified += f"<a href=\"{user_info['url']}\">🔗Link</a>"
-
-    return prettified
-
-
-async def get_github_user(username: str) -> dict | None:
-    url = f"https://api.github.com/users/{username}"
-
-    async with aiohttp.ClientSession() as session:
-        async with session.get(url) as response:
-            if response.status == 404:
-                return None
-
-            j = await response.json()
-            if isinstance(j, dict) and j.get("login") == username:
-                return j
-
-    raise RuntimeError(f"Someting went wrong: {j}")
 
 
 @router.message(Command("github"))
