+++ lab_8/src/utils/__init__.py
@@ -1,0 +1,1 @@
+from .github import get_github_user, prettify_user_info
